Changed paths, works on 1.1.5
I change the path names to use expandPath() and
application.wheels.pluginPath instead of relative paths, that made it
work for me. I also removed the method "$findPluginCfc", it doesn't
seem to be needed, the directory submitted in the form has the
proper-case name already. And just incase someone uses ".CFC" or other
case in their files I fixed changed it from replace to replaceNoCase.